move enkstest to pkg
